Making progress!
This is Local Eats. It still has a long way to go before I call it
finished, as everyday I learn something new I can improve upon it with.
There are also a few ideas I’m hoping to implement in the coming weeks.
This was an ambitious project for me, but I’m proud of how far I’ve
come.